+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Some guidelines in reading this document:
 ## [new release]
 * Fix [#15](https://github.com/log-oscon/redux-wpapi/issues/15) - Support for register dashed-routes as camelCase ([#16](https://github.com/log-oscon/redux-wpapi/pull/16))
 * TTL (Time-to-live) now can be overriden by each request. ([#17](https://github.com/log-oscon/redux-wpapi/pull/17))
+* [#13](https://github.com/log-oscon/redux-wpapi/pull/13)
+  * **[BC]** - `selectQuery` renamed to `selectRequest` and now also accepts `{ cacheID, page }` for selecting Requests
+  * **[BC]** - Ommits `localResources`, now that a denormalization methods exists.
+  * Exposes `selectRequestRaw`, so a request might be selected without denormalization step
+  * Fix [#12](https://github.com/log-oscon/redux-wpapi/issues/12) – Promise returned from middleware dispatch was overlaping responses with same name
+  * Fix [#14](https://github.com/log-oscon/redux-wpapi/issues/14) - lastCacheUpdate should be used as Symbol for preventing collisions, now been tested
 
 ## 1.1.0